Code Cleanup
Following changes were implemented:

-> Formatted the code to look neat
-> Improved the Readme.txt
-> Added a list of TODO list of things to be done.

NOTE:
Update feature will not work as the links are disabled for now. The links will be activated once the release build has been uploaded and then finalized.

Also, the repo has a few files and folders removed so that the formatting and its cleaner now.

Co-Authored-By: DAK <dak404@users.noreply.github.com>